created a reacthook instead of a component
